requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
redis==6.2.0
psycopg[binary]==3.2.9
prometheus-client==0.22.1
python-json-logger==3.3.0
pydantic==2.11.7
grpcio==1.74.0
grpcio-tools==1.74.0
protobuf==6.31.1
numpy==2.3.2
scikit-learn==1.7.1
joblib==1.5.2
kubernetes==32.0.1
PyYAML==6.0.2
pytest==8.4.1
PyJWT==2.10.1
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-httpx==0.57b0
